Extracted methods for UI comparison to AShotService.
