+++ src/main/groovy/com/sysgears/seleniumbundle/core/pagemodel/AbstractPage.groovy
@@ -3,6 +3,7 @@ package com.sysgears.seleniumbundle.core.pagemodel
 import com.codeborne.selenide.*
 import com.sysgears.seleniumbundle.core.conf.Config
 import com.sysgears.seleniumbundle.core.pagemodel.annotations.StaticElement
+import com.sysgears.seleniumbundle.core.uicomparison.AShotService
 import groovy.util.logging.Slf4j
 import org.openqa.selenium.Keys
 
@@ -16,6 +17,11 @@ import static org.testng.Assert.assertTrue
  */
 @Slf4j
 abstract class AbstractPage<T> {
+
+    /**
+     * Instance of AShotService. Used for UI comparison.
+     */
+    protected AShotService aShotService
 
     /**
      * Instance of Config.
@@ -36,6 +42,20 @@ abstract class AbstractPage<T> {
      * URL of the page.
      */
     protected String url
+
+    /**
+     * Captures screenshot and compares it with the previously captured base screenshot. In case there are differences
+     * saves the new screenshot and the image with marked discrepancies. In case there is no previously captured base
+     * screenshot, saves the new screenshot and throws AssertionError.
+     *
+     * @param screenshotName name of a screenshot
+     *
+     * @return instance of the page
+     */
+    T compareLayout(String screenshotName) {
+        aShotService.compareLayout(screenshotName)
+        (T) this
+    }
 
     /**
      * Opens web page.
