+++ src/main/groovy/com/sysgears/seleniumbundle/pagemodel/GooglePage.groovy
@@ -1,16 +1,15 @@
 package com.sysgears.seleniumbundle.pagemodel
 
 import com.codeborne.selenide.SelenideElement
-import com.codeborne.selenide.WebDriverRunner
-import com.sysgears.seleniumbundle.core.pagemodel.AbstractPageWithUIComparison
+import com.sysgears.seleniumbundle.core.pagemodel.AbstractPage
 import com.sysgears.seleniumbundle.core.pagemodel.annotations.StaticElement
-import com.sysgears.seleniumbundle.core.uicomparison.ScreenshotHandler
+import com.sysgears.seleniumbundle.core.uicomparison.AShotService
 import io.qameta.allure.Step
 import org.openqa.selenium.By
 
 import static com.codeborne.selenide.Selenide.$
 
-class GooglePage extends AbstractPageWithUIComparison<GooglePage> {
+class GooglePage extends AbstractPage<GooglePage> {
 
     @StaticElement
     private SelenideElement queryField = $(By.name("q"))
@@ -19,12 +18,8 @@ class GooglePage extends AbstractPageWithUIComparison<GooglePage> {
         this.url = "/"
     }
 
-    GooglePage(String os, String browser) {
-        this.url = "/"
-        this.os = os
-        this.browser = browser
-        handler = new ScreenshotHandler(aShotFactory.getAShotForPage(os, browser, getClass()),
-                WebDriverRunner.getWebDriver())
+    GooglePage(AShotService aShotService) {
+        this.aShotService = aShotService
     }
 
     @Step("Select language")
